+++ Projects/Core/DesignSystem/Demo/Sources/DemoView.swift
@@ -2,22 +2,19 @@ import DesignSystem
 import SwiftUI
 
 public struct DemoView: View {
-    @State var isOn = true
+    @State var isSeleted = true
     @State var selection: Int? = 0
 
     public var body: some View {
         VStack {
             Group {
-                RadioComponent(selection: $selection) {
                     ForEach(0..<4) { index in
                         HStack {
-                            SMSSelectionControls(tag: index)
-                                .radioTag(index)
+                            SMSSelectionControls(isSeleted: $isSeleted)
                             Text("asdfasdf")
                         }
                     }
                 }
-            }
         }
     }
 }
