+++ Projects/Core/DesignSystem/Sources/SelectionControls/SMSSelectionControls.swift
@@ -1,22 +1,42 @@
 import SwiftUI
 
-public struct SMSSelectionControls<T: Hashable>: View {
-    @EnvironmentObject private var value: RadioValue<T>
-    private var tag: T?
+public struct SMSSelectionControls: View {
+    @Binding var isSeleted: Bool
 
-    public init(tag: T?) {
-        self.tag = tag
+    public init(isSeleted: Binding<Bool>) {
+        _isSeleted = isSeleted
+    }
+    public var body: some View {
+        Button {
+            self.isSeleted.toggle()
+        } label: {
+            ZStack {
+                Circle()
+                    .stroke(isSeleted ? Color.sms(.primary(.p2)) : Color.sms(.neutral(.n20)), lineWidth: 2)
+                    .frame(width: 20, height: 20)
+                Circle()
+                    .fill(isSeleted ? Color.sms(.primary(.p2)) : .clear)
+                    .frame(width: 12, height: 12)
+            }
+        }
+        .buttonStyle(CustomButtonStyle(isSeleted: isSeleted))
+    }
+}
+
+private struct CustomButtonStyle: ButtonStyle {
+    var isSeleted: Bool
+
+    init(isSeleted: Bool) {
+        self.isSeleted = isSeleted
     }
 
-    public var body: some View {
-        ZStack {
-            Circle()
-                .stroke(tag == value.selection ? Color.sms(.primary(.p2)) : Color.sms(.neutral(.n20)), lineWidth: 2)
-                .frame(width: 20, height: 20)
-
-            Circle()
-                .fill(tag == value.selection ? Color.sms(.primary(.p2)) : .clear)
-                .frame(width: 12, height: 12)
-        }
+    @ViewBuilder
+    func makeBody(configuration: Configuration) -> some View {
+        Circle()
+            .fill(self.isSeleted && configuration.isPressed ? Color.sms(.neutral(.n20)) : Color.clear)
+            .overlay {
+                configuration.label
+            }
+            .frame(width: 32, height: 32)
     }
 }
